% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkDetectoR.R
\name{get_taxonomy}
\alias{get_taxonomy}
\title{Fetch the entire taxonomy3 table from the API}
\usage{
get_taxonomy()
}
\value{
A data.frame with all rows/columns from taxonomy3.
}
\description{
Fetch the entire taxonomy3 table from the API
}
