% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{iucn_list}
\alias{iucn_list}
\title{Find species at a given coordinate with a specified radius}
\usage{
iucn_list(lat, lon, radius_km = 60)
}
\arguments{
\item{lat}{Numeric, the latitude of the location.}

\item{lon}{Numeric, the longitude of the location.}

\item{radius_km}{Numeric, the radius in kilometers (default is 60).}
}
\value{
A data frame of species present at the location with conditions and probabilities.
}
\description{
This function sends a request to the Flask API, which forwards it to the Plumber API,
to retrieve species data at the specified location and radius.
}
\examples{
iucn_list(43, -70)

}
