% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkDetectoR.R
\name{list_sharks}
\alias{list_sharks}
\title{List Shark Species and Their Performance Metrics}
\usage{
list_sharks()
}
\value{
A data frame with shark species and their performance metrics.
}
\description{
This function reads the CSV file containing classification performance metrics
and lists all shark species along with their precision, recall, and F1 score.
}
