% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkDetectoR.R
\name{performance}
\alias{performance}
\title{Plot Performance Metrics for Genus and Species Classification}
\usage{
performance(genus = "all")
}
\arguments{
\item{genus}{A character string indicating the genus to plot ("all" or a specific genus).}
}
\description{
This function reads the CSV file containing classification performance metrics
and generates bar graphs of precision, recall, and F1 score for genus and species.
When 'genus' is "all", it generates graphs for all genera. When a specific genus is
provided, it generates graphs for that genus and its corresponding species.
}
