% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkDetectorR.R
\name{process_directory}
\alias{process_directory}
\title{Process a Directory of Images for Shark Detection and Classification}
\usage{
process_directory(directory_path = "./images")
}
\arguments{
\item{directory_path}{The path to the directory containing the image files to be processed.}
}
\description{
Processes all images in a specified directory using the Flask API for shark detection and classification.
The API returns detected shark species, detection probabilities, and classified species probabilities for each image.
The results are saved to a CSV file.
}
\examples{
\dontrun{
  process_directory("path/to/images")
}
}
