% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkDetectorR.R
\name{process_image}
\alias{process_image}
\title{Process an Image for Shark Detection and Classification}
\usage{
process_image(image_path)
}
\arguments{
\item{image_path}{The path to the image file to be processed.}
}
\value{
A list with the detected species, detection probability, species probability, and image path.
}
\description{
Submits an image file to the Flask API for processing. The API returns
detected shark species, detection probabilities, and
classified species probabilities along with the processed image path.
Optionally, the function can download the processed image.
}
\examples{
\dontrun{
  result <- process_image("path/to/image.jpg", download_image = FALSE, crop = FALSE)
  print(result)
}
}
