% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkDetectorR.R
\docType{package}
\name{sharkDetectorR-package}
\alias{sharkDetectorR-package}
\alias{sharkDetectorR}
\title{Shark Detection and Classification Package}
\description{
Provides functions to submit videos and images to a Flask API for shark detection and classification.
}
\details{
The package allows users to submit visual media to a backend API, which processes the media with object-detection and image classification models to detect and classify shark species.
}
\keyword{package}
