% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{iucn_list}
\alias{iucn_list}
\title{Find species within a specified bounding box}
\usage{
iucn_list(xmin, ymin, xmax, ymax)
}
\arguments{
\item{xmin}{Numeric, minimum longitude of the bounding box.}

\item{ymin}{Numeric, minimum latitude of the bounding box.}

\item{xmax}{Numeric, maximum longitude of the bounding box.}

\item{ymax}{Numeric, maximum latitude of the bounding box.}
}
\value{
A data frame of species present within the bounding box with conditions and probabilities.
}
\description{
This function sends a request to the Flask API, which forwards it to the Plumber API,
to retrieve species data within the specified bounding box.
}
\examples{
result = iucn_list(-70, 40, -69, 41)

}
