% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R, R/sharkDetectoR.R
\docType{package}
\name{sharkDetectoR-package}
\alias{sharkDetectoR-package}
\alias{sharkDetectoR}
\title{IUCN and Aquamaps Distribution List}
\description{
Provides functions to identify Elasmobranch distributions given a sampling coordinate.

Provides functions to submit videos and images to a Flask API for shark detection and classification.
}
\details{
The package allows users to submit a sampling coordinate and receive a list of Elasmobranch species with overlapping distribution shape files.

The package allows users to submit visual media to a backend API, which processes the media with object-detection and image classification models to detect and classify shark species.
}
\keyword{package}
